+++ com.unity.netcode.gameobjects/Components/NetworkRigidbody2D.cs
@@ -32,7 +32,7 @@ namespace Unity.Netcode.Components
 
         private void FixedUpdate()
         {
-            if (NetworkManager.IsListening)
+            if (IsSpawned)
             {
                 if (HasAuthority != m_IsAuthority)
                 {
@@ -74,7 +74,6 @@ namespace Unity.Netcode.Components
         /// <inheritdoc />
         public override void OnNetworkDespawn()
         {
-            m_IsAuthority = false;
             UpdateRigidbodyKinematicMode();
         }
     }